feat(ui5-checkbox): migrated to Typescript
