+++ packages/main/src/CheckBox.hbs
@@ -1,14 +1,14 @@
 <div
 	class="ui5-checkbox-root {{classes.main}}"
 	role="checkbox"
-	aria-checked="{{ariaChecked}}"
+	aria-checked="{{effectiveAriaChecked}}"
 	aria-readonly="{{ariaReadonly}}"
-	aria-disabled="{{ariaDisabled}}"
+	aria-disabled="{{effectiveAriaDisabled}}"
 	aria-label="{{ariaLabelText}}"
 	aria-labelledby="{{ariaLabelledBy}}"
 	aria-describedby="{{ariaDescribedBy}}"
 	aria-required="{{required}}"
-	tabindex="{{tabIndex}}"
+	tabindex="{{effectiveTabIndex}}"
 	@mousedown="{{_onmousedown}}"
 	@mouseup="{{_onmouseup}}"
 	@keydown="{{_onkeydown}}"
